+++ changelog.md
@@ -26,6 +26,7 @@
 - Added class `Checklists`
 - Added class `Checklists/Items`
 - Added class `ExternalLinkPatterns`
+- Added class `HttpApiModel`
 - **[BREAKING]** Renamed method and class `GraphExecution` to `GraphExecutions`
 - **[BREAKING]** Renamed method and class `Job` to `Jobs`
 - **[BREAKING]** Renamed method and class `Repository` to `Repositories`